refactor(web): moves web directory to electron/renderer again
