+++ src/electron/renderer/components/Screen/index.ts
