+++ src/electron/renderer/components/Sidebar/index.tsx
