+++ src/electron/renderer/components/Sidebar/styles.ts
