+++ src/electron/renderer/features/project/components/CreateModal/styles.ts
