+++ src/electron/renderer/index.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { ThemeProvider } from 'styled-components';
 
-import { theme } from '@/web/styles/theme';
+import { theme } from './styles/theme';
 
-import { App } from '@/web/App';
+import { App } from './App';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
