+++ src/electron/renderer/screens/Nodes/styles.ts
