+++ src/electron/renderer/screens/PortalEntries/index.tsx
