+++ src/electron/renderer/screens/PortalEntries/styles.ts
