+++ src/electron/renderer/screens/PortalsEntities/styles.ts
