+++ src/electron/renderer/screens/Rooms/styles.ts
