+++ src/electron/renderer/styles/colors.ts
