+++ src/electron/renderer/styles/global.ts
