+++ src/electron/renderer/types/styled.d.ts
